Renaming daemontestutils
Clearer, namespaced, although long. But since only used in tests, the long
name won't be an issue.
