+++ windows-agent/cmd/ubuntu-pro-agent/agent/agent_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/canonical/ubuntu-pro-for-wsl/common"
 	"github.com/canonical/ubuntu-pro-for-wsl/windows-agent/cmd/ubuntu-pro-agent/agent"
-	"github.com/canonical/ubuntu-pro-for-wsl/windows-agent/internal/daemon/testutils"
+	"github.com/canonical/ubuntu-pro-for-wsl/windows-agent/internal/daemon/daemontestutils"
 	"github.com/canonical/ubuntu-pro-for-wsl/windows-agent/internal/proservices/registrywatcher/registry"
 	"github.com/stretchr/testify/require"
 )
@@ -511,9 +511,9 @@ func captureStdout(t *testing.T) func() string {
 }
 
 func TestWithWslSystemMock(t *testing.T) {
-	testutils.MockWslSystemCmd(t)
+	daemontestutils.MockWslSystemCmd(t)
 }
 
 func init() {
-	testutils.DefaultNetworkDetectionToMock()
+	daemontestutils.DefaultNetworkDetectionToMock()
 }
