+++ windows-agent/internal/daemon/daemontestutils/networking_mock.go
@@ -1,4 +1,4 @@
-package testutils
+package daemontestutils
 
 import (
 	"errors"
